Move all non-Clojurescript config back to Hydrogen Duct Template
Given that we no longer need to use Duct modules for Cognito or
Keycloak configuration (we have moved it back to the Hydrogen Duct
Template), move statically configured setting to the Template too.

The benefit is that the settings are directly available in
`config.edn` and completely visible, thus having an explicit
configuration instead of an implicit one.