Upgrade TravisCI integration
Among other issues, we were not specifying which JDK version to use
and TravisCI is now defaulting to JDK 11.x, which doesn't work with
our setup. So make sure we specify the right JDK version.

Also there was not automated testing or deployment to Clojars, etc. By
simply using the same TravisCI integration configuration that we are
using in other libraries we fix all those issues at once.