Remove unnecessary preload and closure define
Frankly speaking I don't remember why those deps were even there.
I was almost sure that 10x documentation used to require that in
installation steps but after checking history it's not true.

These dependencies had to be removed because it was crashing the app
whenever we wanted to go to reagent 0.10.x and beyond.