Tune figwheel-main config with ring and css-dirs
Ring server options now provide the desired localost:3449 destination
for figwheel websocket. It's so that the transition from lein-figwheel
is transparent.

The change in css-dirs however is because hydrogen by default
uses duct sass compiler. The scss files in resources are not used
by the index.html. They are only compiled and their product(s)
is put in `target/resources/.../public/css`.