Change namespace for ClojureScript main entry point
The goal is to be able to move all the ClojureScrpt files inside a
single directory. This way we can configure the Figwheel-main watcher
to watch just that directory, instead of the top lever `src` directory
that uses by default. This avoids triggering rebuilds on Clojure files
changes.

In other to achive it, we need to move the ClojureScript main entry
point namespace inside that directory. But that changes the name of
the symbol that we need to pass to both Figwheel-main and the Google
Closure compiler.